refactor(move): move the code to process CLI from `src/index.ts` to `src/cli.ts`
